made all of vue be precompiled
